# Fix critical performance issues: search lag and audio playback delay